server/requirements.txt
fastapi
psycopg2-binary
uvicorn
passlib[bcrypt]
python-jose
python-dotenv